Consolidated App.jsx, Homepage.jsx and completed the About write-up
