+++ client/src/pages/About.jsx
@@ -7,6 +7,7 @@ function About() {
         <>
         <CowPatternBg/>
             <h1>About Page</h1>
+            <p>Welcome to our app, a safe haven for your thoughts and emotions. In a world full of noise, take a moment to find your voice here. Let your feelings flow freely onto the pages of your digital journal. Whether it's joy, anger, sadness, or confusion, each entry is a step towards understanding and healing.<br />We understand the weight that burdens the mind. That's why we're here to offer a listening ear, a virtual shoulder to lean on. Our resources are here not just to guide, but to accompany you on your journey towards mental well-being. You're not alone in this struggle, and together, we can navigate the twist and turns of life's challenges.<br /> Take the first step towards a lighter heart and a clearer ming.<br /> START YOUR JOURNEY WITH US TODAY.</p>
         </>
     )
 }
