+++ client/src/pages/Homepage.jsx
@@ -2,8 +2,7 @@ import React from 'react';
 import {NavLink} from 'react-router-dom'
 
 import { useState } from 'react'
-import CowPatternBg from '../components/CowPatternBg';
-import AddNewEntry from '../components/AddNewEntry'
+import {CowPatternBg, AddNewEntry } from '../components';
 import '../styles/pages/homepage.scss'
 
 import {useStore} from '../store'
